Move LoggerModule to di package and add TestLoggerModule
