+++ app/src/main/kotlin/at/bitfire/davdroid/di/LoggerModule.kt
@@ -2,7 +2,7 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.davdroid.log
+package at.bitfire.davdroid.di
 
 import dagger.Module
 import dagger.Provides
